+++ cmd/commander/commander.go
@@ -53,6 +53,15 @@ Tests are always executed in alphabetical order.
 
 Examples:
 
+Directory test:
+commander test --dir /your/dir/
+
+Stdin test:
+cat commander.yaml | commander test -
+
+HTTP test:
+test https://your-url/commander_test.yaml
+
 Filtering tests:
 test commander.yaml --filter="my test"